Move action-discovery methods to RedfishManager for non-Dell vendors
discover_redfish_actions, discover_member_redfish_actions and _get_actions were Dell-only, but invoke_action (already on RedfishManager) calls discover_redfish_actions -- so the Supermicro exporter AttributeError'd on a non-Dell instance. All three are vendor-neutral (standard Actions/Members/Oem keys, no Dell logic), so move them to RedfishManager (REDFISH_JSON -> RedfishJson; RedfishAction promoted from TYPE_CHECKING to a runtime import). IDracManager inherits them via MRO.

Structurally verified (the AST move ran ast.parse on both files before writing); the runtime MRO/import check is blocked by the intermittent Python-startup hang, so it must run on CI/gb300.
